pyproject.toml
[project]
name = "django-weather-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.4",
    "django-filter>=24.3",
    "django-cotton>=1.5.1",
    "django-ninja>=1.3.0",
    "colorlog>=6.9.0",
    "aiohttp>=3.11.11",
    "daphne>=4.1.2",
    "channels>=4.2.0",
    "requests>=2.32.3",
    "django-htmx>=1.21.0",
    "python-dotenv>=1.0.1",
    "whitenoise>=6.8.2",
    "django-debug-toolbar>=4.4.6",
    "aiocache[memcached]>=0.12.3",
    "django-compressor>=4.5.1",
    "django-minify-html>=1.11.0",
    "pytailwindcss>=0.2.0",
    "pymemcache>=4.0.0",
]
